// src/app/inventario/interfaces/boxes.interfaces.ts
// Generated by https://quicktype.io

export interface Boxes {
  status: boolean;
  cajas:  Caja[];
}

export interface GetBox {
  status:  boolean;
  message: string;
  caja:    Caja;
}

export interface Caja {
  id:          number;
  barcode:     string;
  stock_cajas: number;
  stock_tapas: number;
  producto_id: number;
  created_at:  string;
  updated_at:  string;
}